# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "IAM API REST in Golang with Gin Framework",
        "title": "Lab Routing Service with golang",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/actions": {
            "get": {
                "description": "Get available action from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Get Available Actions",
                "operationId": "get-action",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Action"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "int"
                        }
                    }
                }
            },
            "post": {
                "description": "Post new action to DB",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actions"
                ],
                "summary": "Post new action",
                "operationId": "post-action",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Action"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "int"
                        }
                    }
                }
            }
        },
        "/isActive": {
            "get": {
                "description": "Get deployed version of the api from db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sample"
                ],
                "summary": "Get Current deployed version",
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Action": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string"
                }
            }
        }
    }
}